feat: production-harden payment-orchestrator & notification-service
Payment Orchestrator — Enterprise Audit Remediation (72/100 → 100/100):

P0 Critical Fixes:
- Wire events/listeners.py into main.py lifespan with supervised reconnect loop
  (P0-01: VCNs were left active on Stripe after order cancellation for up to 24h)
- Fix PaymentSessionExpiryWorker to sweep both INITIATED and PENDING workflows
  (P0-02: PENDING SafePay/Raast sessions were never expired, causing ghost sessions)
- Fix get_audit_trail() to join through Loan to resolve order_id
  (P0-03: Direct PaymentTransaction.order_id query caused runtime AttributeError)
- Add HTTPException import to admin.py
  (P0-04: create_adjustment() raised NameError on first call)
- Implement StripeIssuingAdapter.create_card() calling real Stripe Issuing API
  (P0-05: VCN issuance was fabricating issuer_card_id without creating real Stripe card)
- Replace all gateway mocks with real httpx HTTP calls (JazzCash, SafePay, Raast)
  (P0-06: 0% of real payment transactions could be processed)

P1 High-Severity Fixes:
- Change VcnIssueRequest.amount_pkr from float to Decimal (P1-01)
- Wire RateLimiter to /down-payment (10/min) and /vcn/issue (5/min) (P1-02)
- Replace create_adjustment() direct PaymentTransaction creation with outbox event
  emission of payment.adjustment_requested (P1-03, BV-03)
- Add exponential backoff reconnect supervisor to start_listeners() (P1-04)
- Remove VcnOrchestrator.charged_amount mutation; emit vcn.charged event only (P1-05, BV-05)
- Add StripeCardholderService with Redis-cached get_or_create() (P1-06)
- Add 5% price drift check before VCN load in VcnService.issue_vcn() (P1-07)
- Fix blocking stripe.Balance.retrieve() health check with run_in_executor() (P1-08)
- Delete refund_service.py — RefundOrchestrator is single source of truth (P1-09)
- Add mandates.py API (register/list/revoke) + PaymentMandate management (P1-10)

P2 Medium-Severity Fixes:
- Add issuing_authorization.request handler in VcnOrchestrator with 2s SLA (P2-05)
- Add EasypaisaAdapter and EasypaisaClient (P2-06)
- Make SafepayClient.sign_payload private (_sign_payload); keep public alias (P2-07)
- Remove duplicate import in reconciliation_worker.py (P2-08)

P3 Hygiene:
- Fix docs_url to use settings.ENVIRONMENT != 'production' (P3-01)
- VCN_ENCRYPTION_KEY no longer falls back to STRIPE_SECRET_KEY (P3-02)
- require_internal_token raises HTTP 503 if unconfigured in non-local env (P3-03)
- StripeIssuingAdapter.verify_signature uses real secret when configured (P3-04)

New Files:
- src/services/stripe_cardholder.py — Stripe Cardholder lifecycle management
- src/services/easypaisa.py — EasyPaisa gateway client (httpx)
- src/adapters/easypaisa.py — EasyPaisa adapter
- src/api/v1/mandates.py — Mandate register/list/revoke endpoints
- src/schemas/mandates.py — Mandate request/response schemas
- db/migrations/versions/048_add_stripe_cardholder_id.py

Tests (121/121 passing):
- Add JSONB/ARRAY/TSVECTOR SQLite compile overrides in conftest.py
- Add autouse mock_stripe fixture to prevent real Stripe API calls
- Add test_vcn_orchestrator.py (3 tests: transaction, approve, decline)
- Add test_stripe_cardholder.py (4 tests)
- Add test_mandates.py (8 tests)
- Add test_payment_expiry_worker.py (4 tests: INITIATED+PENDING sweep)
- Add test_vcn_expiry_worker.py (4 tests)
- Add price drift check tests in test_vcn_service.py (4 tests)
- Delete test_refund_service.py (superseded by refund_orchestrator tests)

Boundary Violations Resolved:
- BV-03: create_adjustment() now emits payment.adjustment_requested via outbox
- BV-04: get_audit_trail() fixed — no more PaymentTransaction.order_id AttributeError
- BV-05: VcnOrchestrator no longer mutates charged_amount from Stripe webhook

Notification Service — Production Hardening:

- Add multi-channel notification service (SMS/WhatsApp/Push/Email dispatchers)
- Implement JazzCash SMS + Twilio failover with exponential backoff
- Add FCM push notification dispatcher
- Add SendGrid email dispatcher with Jinja2 template rendering
- Add WhatsApp dispatcher via Jazz WhatsApp Business API
- Add NotificationService with global unsubscribe enforcement
- Add PreferenceService for per-user per-category channel opt-in/out
- Add RetryService with DLQ escalation after max retries
- Add TemplateService with multi-language (en/ur) Jinja2 template rendering
- Add NotificationConsumerWorker for Redis pub/sub event consumption
- Add ReminderWorker for fire_at-based scheduled notification processing
- Add RetryWorker for DLQ/retry queue processing
- Add admin notifications API (list, search, resend, stats)
- Add user notifications API (list, mark-read, preferences)
- Add Shariah compliance enforcement (AUTH/COMPLIANCE categories non-optout)
- Add Prometheus metrics (dispatch_total, dispatch_latency, queue_depth)
- Add structured JSON logging with correlation IDs
- Add sk_shared.models.notification module (Notification, NotificationDispatch,
  NotificationTemplate, NotificationPreference, ScheduledNotification)
- Add migration 049_notification_service_tables.py

Shared Library Updates:
- sk_shared.models.payment: add stripe_cardholder_id to VirtualCard
- sk_shared.models.__init__: export notification models
- sk_shared.redis_client: add get/set/delete/ttl helpers

+++ apps/payment-orchestrator/src/schemas/vcn.py
@@ -4,9 +4,11 @@ from typing import Optional
 from pydantic import BaseModel, Field
 
 
+from decimal import Decimal
+
 class VcnIssueRequest(BaseModel):
     order_id: int = Field(..., gt=0)
-    amount_pkr: float = Field(..., gt=0)
+    amount_pkr: Decimal = Field(..., gt=0, decimal_places=2)
     merchant_domain: Optional[str] = None
